Check if there is room for tree to grow
Previously tree saplings would grow regardless of the blocks above it.
This could be abused to destroy bedrock and other nodes.

To solve this, checks are added for nodes around the trees.  The volume
checked depends on the tree type.  The volume for each tree has been set
so reject trees where branches would grow into other nodes.  Some nodes
types like torches, plants and wood are ignored in the checks.

Saplings in a 2x2 formation will grow into normal trees if there is no
room for a huge tree.  Oak trees has separate checks for balloon oak
trees and normal oak trees, if there is no room for a balloon oak, it
becomes a normal oak tree.