Fixes #2976
Now the inventory view on creative is no longer set to the survival
inventory upon taking damage.
Also the initial inventory view when openning a world in creative is at
the search page, not the survival inventory page.

mcl_inventory/init.lua no longer wraps mcl_armor.update_player to also
update the inventory view. This visual update happens anyways since it
is registered to happen on a visual change with
mcl_player.register_on_visual_change .

This also fixes a small visual bug that, on creative mode, if you would
equip/unequip an armor piece, the image of the player would be updated
twice.