Avoid multiple ruined portals in the same chunk
This reduces the fill rate from 38 to 1 spawn attempts, and to counter this effect increases the chunk probability by a factor of 40.

This should yield roughly the same amount of portals, but never produce two in the same map generation chunk. Before, they would frequently cluster. (Although out of the 38 attempts, apparently only 0-3 are usually successful).